1) Added support for new camera mode: end camera positions and types    can now be specified in the scene file. Currently supported are:    - normal 'ahead of kart' camera (which is what we used so far)    - fixed camera: camera is at a fixed position, always looking      at the kart. 2) Removed camera.hpp dependency from kart.hpp.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@5585 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/utils/string_utils.cpp
@@ -360,7 +360,7 @@ namespace StringUtils
      */
     std::string timeToString(float time)
     {
-        int int_time   = time*100.0f+0.5f;
+        int int_time   = (int)(time*100.0f+0.5f);
         int min        = int_time / 6000;
         int sec        = (int_time-min*6000)/100;
         int hundredths = (int_time - min*6000-sec*100);
